Move optimizer into optim.go
