+++ torch.go
@@ -7,8 +7,6 @@ package gotorch
 import "C"
 import (
 	"reflect"
-	"runtime"
-	"unsafe"
 )
 
 // Optimizer struct
@@ -25,43 +23,6 @@ func RandN(rows, cols int, requireGrad bool) Tensor {
 	t := C.RandN(C.int(rows), C.int(cols), C.int(rg))
 	setTensorFinalizer(&t)
 	return Tensor{&t}
-}
-
-// NewSGDOpt creates a SGD Optimizer
-func NewSGDOpt(lr, momentum, dampening, weightDecay float64, nesterov bool) Optimizer {
-	nt := 0
-	if nesterov {
-		nt = 1
-	}
-	sgd := C.SGD(C.double(lr), C.double(momentum), C.double(dampening),
-		C.double(weightDecay), C.int(nt))
-	runtime.SetFinalizer(&sgd, func(p *C.Optimizer) { C.Optimizer_Close(*p) })
-	return Optimizer{&sgd}
-}
-
-// AddParameters adds parameters
-func (opt Optimizer) AddParameters(tensors []Tensor) {
-	CT := []unsafe.Pointer{}
-	for _, t := range tensors {
-		CT = append(CT, unsafe.Pointer(*t.T))
-	}
-	p := (*reflect.SliceHeader)(unsafe.Pointer(&CT)).Data
-	C.AddParameters(*opt.Opt, (*C.Tensor)(unsafe.Pointer(p)), C.int(len(CT)))
-}
-
-// ZeroGrad reset gradients to zero
-func (opt Optimizer) ZeroGrad() {
-	C.ZeroGrad(*opt.Opt)
-}
-
-// Step updates parameters
-func (opt Optimizer) Step() {
-	C.Step(*opt.Opt)
-}
-
-// Close the optimizer
-func (opt Optimizer) Close() {
-	C.Optimizer_Close(*opt.Opt)
 }
 
 // Module interface
